requirements.txt
astroid==2.5.6
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
Django==2.2.24
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.8.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
wrapt==1.12.1
